feat：Move `FeedImageViewModel` to separate file
