+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedImagePresenter.swift
@@ -7,18 +7,6 @@
 
 import Foundation
 
-public struct FeedImageViewModel<Image> {
-    public let description: String?
-    public let location: String?
-    public let image: Image?
-    public let isLoading: Bool
-    public let shouldRetry: Bool
-
-    public var hasLocation: Bool {
-        return location != nil
-    }
-}
-
 public protocol FeedImageView {
     associatedtype Image
